Create shared
A shared module was created, containing:

Common TypeScript types for frontend and backend
Helper functions for data validation and normalization
TypeScript configuration for ES modules
The backend was reorganized according to the MVC pattern:

Controllers supporting business logic (chat, moderation, cache)
Routers defining API endpoints
Common functionality was separated into a shared module
Updated the documentation in README.md, adding a detailed description of the new structure.
