+++ packages/backend/src/controllers/index.mjs
@@ -1,0 +1,3 @@
+export { ChatController } from "./chatController.mjs";
+export { ModerationController } from "./moderationController.mjs";
+export { CacheController } from "./cacheController.mjs";
