+++ packages/backend/src/routes/index.mjs
@@ -1,0 +1,3 @@
+export { createChatRouter } from "./chatRoutes.mjs";
+export { createModerationRouter } from "./moderationRoutes.mjs";
+export { createCacheRouter } from "./cacheRoutes.mjs";
