+++ packages/shared/src/index.ts
@@ -1,2 +1,2 @@
-export * from "./types";
-export * from "./utils";
+export * from "./types/chat.js";
+export * from "./utils/validation.js";
